renamed navigation components and updated explore page based on new names
