--- client/src/components/Header/index.css
@@ -1,1 +1,0 @@
-
